backend/distribution_manager/requirements.txt
kafka-python==2.0.2
twilio==7.4.0
